[WIP] 1503 Refactor Asset Icon Component
