+++ packages/components/src/local/map-icon/README.md
@@ -1,0 +1,27 @@
+# Asset Icon documentation
+
+This is an example readme file for the 'Asset Icon' component.
+
+## UI Example
+
+<!-- STORY -->
+
+## Code example
+
+<!-- SOURCE -->
+
+## Props
+
+<!-- PROPS -->
+
+## Usage
+
+### ES6 Import
+```js
+import AssetIcon from './asset-icon'
+```
+
+### CommonJS
+
+```js
+const AssetIcon = require('./asset-icon')
