+++ packages/components/src/local/map-icon/styles.module.scss
@@ -1,0 +1,121 @@
+.asset-icon-background {
+  position: relative;
+  border-radius: 100%;
+  width: 100%;
+  height: 100%;
+}
+
+.asset-icon-fw {
+  width: 100%;
+  height: 100%;
+}
+
+.asset-icon-invert {
+  filter: invert(100%);
+}
+
+.asset-icon {
+  position: absolute;
+  border-radius: 50%;
+  background-repeat: no-repeat;
+  background-position: 50%;
+  background-size: 80%;
+  box-sizing: border-box;
+  width: 100%;
+  height: 100%;
+  border: 0;
+  padding: 10%;
+  &.selected {
+    border: 4px solid rgba(255, 255, 255, 0.7);
+    box-shadow: 3px 3px 2px rgba(0, 0, 0, 0.25);
+    border-radius: 10%;
+  }
+  &.destroyed {
+    border-radius: 40%;
+    filter: brightness(0.7);
+  }
+}
+
+.force {
+  text-align: center;
+  padding: 1rem;
+}
+
+.blue {
+  background-color: #000fff;
+}
+
+.red {
+  background-color: #ff0b00;
+}
+
+.blue-1 {
+  background-color: #0af;
+}
+
+.red-1 {
+  background-color: #f00;
+}
+
+.blue-2 {
+  background-color: #00008f;
+}
+
+.red-2 {
+  background-color: #ff7b7b;
+}
+
+.green {
+  background-color: #19bd37;
+}
+
+.unknown {
+  background-color: #999;
+}
+
+.popup {
+  position: absolute;
+  top: 100px;
+  left: 100px;
+  z-index: 999;
+}
+
+.orientation {
+  /* stylelint-disable */
+  -webkit-mask: url("./counters/orientation-marker.svg");
+  /* stylelint-enable */
+  mask: url("./counters/orientation-marker.svg");
+}
+
+.triangle {
+  fill-opacity: 0.6;
+}
+
+.img {
+  width: 100%;
+  height: 100%;
+  background-repeat: no-repeat;
+  background-position: 50%;
+  background-size: 100%;
+}
+
+.orient-icon-with-image {
+  padding: 10%;
+  box-sizing: border-box;
+  background-color: transparent;
+  img {
+    width: 100%;
+    display: block;
+  }
+}
+
+.asset-icon-with-image {
+  box-sizing: border-box;
+  width: 100%;
+  height: 100%;
+
+  img {
+    width: 100%;
+    display: block;
+  }
+}
