+++ packages/components/src/local/map-icon/styles.module.scss.d.ts
@@ -1,0 +1,27 @@
+// This file is automatically generated.
+// Please do not change this file!
+interface CssExports {
+  'asset-icon': string;
+  'asset-icon-background': string;
+  'asset-icon-fw': string;
+  'asset-icon-invert': string;
+  'asset-icon-with-image': string;
+  'blue': string;
+  'blue-1': string;
+  'blue-2': string;
+  'destroyed': string;
+  'force': string;
+  'green': string;
+  'img': string;
+  'orient-icon-with-image': string;
+  'orientation': string;
+  'popup': string;
+  'red': string;
+  'red-1': string;
+  'red-2': string;
+  'selected': string;
+  'triangle': string;
+  'unknown': string;
+}
+export const cssExports: CssExports;
+export default cssExports;
